Move finalize_data to common and add to all pipelines
