+++ src/tests/unit/nodes/test_betting.py
@@ -56,15 +56,3 @@ class TestBetting(TestCase, ColumnAssertionMixin):
             valid_data_frame=valid_data_frame,
             feature_function=feature_function,
         )
-
-    def test_finalize_data(self):
-        data_frame = (
-            fake_cleaned_match_data(N_MATCHES_PER_SEASON, YEAR_RANGE)
-            .assign(nans=None)
-            .astype({"year": "str"})
-        )
-
-        finalized_data_frame = betting.finalize_data(data_frame)
-
-        self.assertEqual(finalized_data_frame["year"].dtype, int)
-        self.assertFalse(finalized_data_frame["nans"].isna().any())
